Storage tests

+++ src/Tests/Gaucho.Integration.Tests/EventBusLoadTests.cs
@@ -56,7 +56,7 @@ namespace Gaucho.Test.LoadTests
             ProcessingServer.Server.WaitAll(pipelineId);
 
 			var storage = GlobalConfiguration.Configuration.Resolve<IStorage>();
-			Assert.GreaterOrEqual(100 * 10, storage.Get<int>(pipelineId, "ProcessedEventsMetric"));
+			Assert.GreaterOrEqual(100 * 10, storage.Get<long>(pipelineId, "ProcessedEventsMetric"));
 
 			var stats = new StatisticsApi(pipelineId);
 			Assert.AreEqual(1, stats.GetMetricValue(MetricType.ThreadCount));
